use HumanizerInterface to allow easier humanization override
